% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_clean_before_linkage.R
\name{all_clean_before_linkage}
\alias{all_clean_before_linkage}
\title{Clean data in preparation for linkage}
\usage{
all_clean_before_linkage(
  baby_to_match,
  parents_to_match,
  key_baby,
  type_key_baby,
  key_parent,
  type_key_parent
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{key_baby}{variable name(s) to match on from \code{baby_to_match}.}

\item{type_key_baby}{Character string. The variable type of \code{key_baby}, used to
determine which cleaning function to apply. One of \code{"name"} or \code{"address"}. For other types, use \code{""}.}

\item{key_parent}{variable name(s) to match on from \code{parents_to_match}.}

\item{type_key_parent}{Character string. The variable type of \code{key_parent}, used to
determine which cleaning function to apply. One of \code{"name"} or \code{"address"}. For other types, use \code{""}.}
}
\value{
A list containing cleaned versions of \code{baby_to_match} and \code{parents_to_match}.
}
\description{
Clean data in preparation for linkage
}
