% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_candidate_parent_pool.R
\name{build_candidate_parent_pool}
\alias{build_candidate_parent_pool}
\title{Build pool of candidate parent matches based on multiple variables}
\usage{
build_candidate_parent_pool(
  baby_to_match,
  parents_to_match,
  key_baby,
  key_parent,
  id_parent,
  jaro_col_name
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{key_baby}{string variable name(s) to match on from \code{baby_to_match}.}

\item{key_parent}{string variable name(s) to match on from \code{parents_to_match}.}

\item{id_parent}{string variable name to identify observations in \code{parents_to_match}.}

\item{jaro_col_name}{string variable name(s) for the added column of the Jaro similarity corresponding to the keys.}
}
\value{
A vector containing all unique \code{id_parent} values matching one or more keys.
}
\description{
Build pool of candidate parent matches based on multiple variables
}
