% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_baby_parent_linkage.R
\name{combined_baby_parent_linkage}
\alias{combined_baby_parent_linkage}
\title{Combined deterministic + probabilistic parent-child linkage based on Jaro similarity using a single matching key}
\usage{
combined_baby_parent_linkage(
  baby_to_match,
  parents_to_match,
  key_baby,
  key_parent,
  jaro_col_name = "jaro_similarity"
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{key_baby}{string variable name to match on from \code{baby_to_match}.}

\item{key_parent}{string variable name to match on from \code{parents_to_match}.}

\item{jaro_col_name}{if needed, string variable name for the added column of the Jaro similarity. Default is \code{jaro_col_name = "jaro_similarity"}.}
}
\value{
A list containing
\describe{
\item{\code{match}}{A copy of \code{baby_to_match} with added columns for
the best-matching parent.}
\item{\code{type}}{A string indicator of the type of linkage used:
\code{"deterministic"} or \code{"probabilistic"}.}
}
}
\description{
Combined deterministic + probabilistic parent-child linkage based on Jaro similarity using a single matching key
}
