% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_baby_parent_linkage.R
\name{deterministic_baby_parent_linkage}
\alias{deterministic_baby_parent_linkage}
\title{Deterministic parent-child linkage based on exact matches using a single matching key}
\usage{
deterministic_baby_parent_linkage(
  baby_to_match,
  parents_to_match,
  key_baby,
  key_parent
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{key_baby}{string variable name to match on from \code{baby_to_match}.}

\item{key_parent}{string variable name to match on from \code{parents_to_match}.}
}
\value{
A copy of \code{baby_to_match} with added columns for the exact-matching parent(s) (if applicable).
}
\description{
Deterministic parent-child linkage based on exact matches using a single matching key
}
