% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_candidate_parent_pool.R
\name{evaluate_candidate_parent_pool}
\alias{evaluate_candidate_parent_pool}
\title{Evaluate pool of candidate parent matches based on multiple variables to identify "best" match}
\usage{
evaluate_candidate_parent_pool(
  baby_to_match,
  parents_to_match,
  key_baby,
  key_parent,
  id_parent,
  jaro_col_name,
  candidate_pool = NULL,
  aggregate_scores = "mean"
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{key_baby}{string variable name(s) to match on from \code{baby_to_match}.}

\item{key_parent}{string variable name(s) to match on from \code{parents_to_match}.}

\item{id_parent}{string variable name to identify observations in \code{parents_to_match}.}

\item{jaro_col_name}{string variable name(s) for the added column of the Jaro similarity corresponding to the keys.}

\item{candidate_pool}{optional, vector of \code{id_parent} values returned by the \code{build_candidate_parent_pool} function.}

\item{aggregate_scores}{string for how scores should be aggregated across keys. Default is \code{aggregate_scores = "mean"}; other options are \code{"median"} and \code{"sum"}.}
}
\value{
A list containing
\describe{
\item{\code{best_candidate}}{A copy of the best-matching candidate from \code{parents_to_match} with added columns for
the similarity scores per variable and aggregate.}
\item{\code{all_candidates}}{A copy of alll candidates from \code{parents_to_match} with added columns for
the similarity scores per variable and aggregate.}
}
}
\description{
Evaluate pool of candidate parent matches based on multiple variables to identify "best" match
}
