% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaro_baby_parent_linkage.R
\name{jaro_baby_parent_linkage}
\alias{jaro_baby_parent_linkage}
\title{Probabilistic parent-child linkage based on Jaro similarity using a single matching key}
\usage{
jaro_baby_parent_linkage(
  baby_to_match,
  key_baby,
  key_parent,
  jaro_col_name = "jaro_similarity"
)
}
\arguments{
\item{baby_to_match}{data.frame (with one row) for a single baby that includes at least \code{key_baby} as a column.}

\item{key_baby}{variable name to match on from \code{baby_to_match}.}

\item{key_parent}{variable name to match on from \code{parents_to_match}.}

\item{parents_to_match}{data.frame (with multiple rows) for all potential parent matches that includes at least \code{key_parent} as a column.}

\item{name_jaro_col}{variable name for the added column of the Jaro similarity. Default is \code{name_jaro_col = "jaro_similarity"}.}
}
\value{
A copy of \code{baby_to_match} with added columns for the best-matching parent.
}
\description{
Probabilistic parent-child linkage based on Jaro similarity using a single matching key
}
